Correcting defects in the final design.
